Updated to 15.11.04.001
This is such a pain to update. For future reference: You need to
replace the invisible files with the new invisible files. To do this
you must change their names (to 1.txt, for example) as the . before
their names causes them to be invisible. Then you copy and replace as
normal. Next you need to get the opmodes back. They are stored in
ftc_app_master/FtcRobotController/src/main/java/com/qualcomm/ftcrobotcon
troller/opmodes/